ajout condition pour ne pas pouvoir modifier un item déjà archivé

+++ src/main/java/com/ebock/service/ItemService.java
@@ -185,44 +185,47 @@ public class ItemService {
             throw new ForbiddenException("Not your item");
         }
 
-        if(item.quantity == 0){
-            itemMapper.archiveRoomsById(item.itemId);
-        }
+        if(existingItem.quantity != 0) {
 
-        itemMapper.update(cip, item);
+            if (item.quantity == 0) {
+                itemMapper.archiveRoomsById(item.itemId);
+            }
 
-        // Update tags
-        if (itemUpdatePayload.tagList != null) {
-            itemTagMapper.deleteByItemId(itemId);
-            if (!itemUpdatePayload.tagList.isEmpty()) {
-                itemTagMapper.insert(item.itemId, itemUpdatePayload.tagList);
+            itemMapper.update(cip, item);
+
+            // Update tags
+            if (itemUpdatePayload.tagList != null) {
+                itemTagMapper.deleteByItemId(itemId);
+                if (!itemUpdatePayload.tagList.isEmpty()) {
+                    itemTagMapper.insert(item.itemId, itemUpdatePayload.tagList);
+                }
+            }
+
+            // Update images
+            if (itemUpdatePayload.imageList != null) {
+                itemImageMapper.deleteByItemId(itemId);
+                if (!itemUpdatePayload.imageList.isEmpty()) {
+                    itemImageMapper.insert(item.itemId, itemUpdatePayload.imageList);
+                }
+            }
+
+            // Update payment option
+            if (itemUpdatePayload.paymentOptionList != null) {
+                itemPaymentOptionMapper.deleteByItemId(itemId);
+                if (!itemUpdatePayload.paymentOptionList.isEmpty()) {
+                    itemPaymentOptionMapper.insert(item.itemId, itemUpdatePayload.paymentOptionList);
+                }
+            }
+
+            // Update delivery option
+            if (itemUpdatePayload.deliveryOptionList != null) {
+                itemDeliveryOptionMapper.deleteByItemId(itemId);
+                if (!itemUpdatePayload.deliveryOptionList.isEmpty()) {
+                    itemDeliveryOptionMapper.insert(item.itemId, itemUpdatePayload.deliveryOptionList);
+                }
             }
         }
-
-        // Update images
-        if (itemUpdatePayload.imageList != null) {
-            itemImageMapper.deleteByItemId(itemId);
-            if (!itemUpdatePayload.imageList.isEmpty()) {
-                itemImageMapper.insert(item.itemId, itemUpdatePayload.imageList);
-            }
-        }
-
-        // Update payment option
-        if (itemUpdatePayload.paymentOptionList != null) {
-            itemPaymentOptionMapper.deleteByItemId(itemId);
-            if (!itemUpdatePayload.paymentOptionList.isEmpty()) {
-                itemPaymentOptionMapper.insert(item.itemId, itemUpdatePayload.paymentOptionList);
-            }
-        }
-
-        // Update delivery option
-        if (itemUpdatePayload.deliveryOptionList != null) {
-            itemDeliveryOptionMapper.deleteByItemId(itemId);
-            if (!itemUpdatePayload.deliveryOptionList.isEmpty()) {
-                itemDeliveryOptionMapper.insert(item.itemId, itemUpdatePayload.deliveryOptionList);
-            }
-        }
-
+        
         return Response.noContent().build();
     }
